requirements.txt
pytest==6.1.2
pytest-asyncio==0.14.0
pytest-cov==2.10.1
flake8==3.8.4
PyJWT~=2.3.0
isort==5.9.3
black==20.8b1
